[WIP] Detections: choose smaller boxes: update default front

+++ Editor/Controls/MapEditor/controlmapeditor-raycasting.cpp
@@ -449,9 +449,9 @@ void ControlMapEditor::getCorrectPositionOnRay(Position &position,
     position.setZ(z);
     if (!square)
     {
-        position.setCenterX(static_cast<float>(qRound(point.x() + 1) %
+        position.setCenterX(static_cast<double>(qRound(point.x() + 1) %
             m_map->squareSize()) / m_map->squareSize() * 100);
-        position.setCenterZ(static_cast<float>(qRound(point.z() + 1) %
+        position.setCenterZ(static_cast<double>(qRound(point.z() + 1) %
             m_map->squareSize()) / m_map->squareSize() * 100);
     } else
     {
